+++ logprep/processor/field_manager/processor.py
@@ -29,15 +29,13 @@ Processor Configuration
 .. automodule:: logprep.processor.field_manager.rule
 """
 
-import itertools
 from collections import namedtuple
 
 from logprep.abc.processor import Processor
-from logprep.processor.base.exceptions import FieldExistsWarning
 from logprep.processor.field_manager.rule import FieldManagerRule
 from logprep.util.helper import (
     add_and_overwrite,
-    add_field_to,
+    add_fields_to,
     get_dotted_field_value,
     pop_dotted_field_value,
 )
@@ -72,34 +70,23 @@ class FieldManager(Processor):
         self._write_to_single_target(args, extend_target_list, overwrite_target, rule)
 
     def _apply_mapping(self, event, rule, rule_args):
-        source_fields, _, mapping, _, _ = rule_args
+        source_fields, _, mapping, extend_target_list, overwrite_target = rule_args
         source_fields, targets = list(zip(*mapping.items()))
         source_field_values = self._get_field_values(event, mapping.keys())
         self._handle_missing_fields(event, rule, source_fields, source_field_values)
         if not any(source_field_values):
             return
         source_field_values, targets = self._filter_missing_fields(source_field_values, targets)
-        self._write_to_multiple_targets(event, targets, source_field_values, rule, rule_args)
+        add_fields_to(
+            event,
+            dict(zip(targets, source_field_values)),
+            rule,
+            extend_target_list,
+            overwrite_target,
+        )
         if rule.delete_source_fields:
             for dotted_field in source_fields:
                 pop_dotted_field_value(event, dotted_field)
-
-    def _write_to_multiple_targets(self, event, target_fields, field_values, rule, rule_args):
-        _, _, _, extend_target_list, overwrite_target = rule_args
-        results = map(
-            add_field_to,
-            itertools.repeat(event, len(target_fields)),
-            target_fields,
-            field_values,
-            itertools.repeat(extend_target_list, len(target_fields)),
-            itertools.repeat(overwrite_target, len(target_fields)),
-        )
-        if not all(results):
-            unsuccessful_indices = [i for i, x in enumerate(results) if not x]
-            unsuccessful_targets = [
-                x for i, x in enumerate(target_fields) if i in unsuccessful_indices
-            ]
-            raise FieldExistsWarning(rule, event, unsuccessful_targets)
 
     def _write_to_single_target(self, args, extend_target_list, overwrite_target, rule):
         event, target_field, source_fields_values = args
@@ -122,7 +109,7 @@ class FieldManager(Processor):
             case State(
                 extend=True, overwrite=True, single_source_element=False, target_is_list=False
             ):
-                add_and_overwrite(event, target_field, source_fields_values)
+                add_and_overwrite(event, fields={target_field: source_fields_values}, rule=rule)
                 return
 
             case State(
@@ -134,16 +121,16 @@ class FieldManager(Processor):
             ):
                 flattened_source_fields = self._overwrite_from_source_values(source_fields_values)
                 source_fields_values = [*flattened_source_fields]
-                add_and_overwrite(event, target_field, source_fields_values)
+                add_and_overwrite(event, fields={target_field: source_fields_values}, rule=rule)
                 return
 
             case State(extend=True, overwrite=False, target_is_list=False, target_is_none=True):
-                add_and_overwrite(event, target_field, source_fields_values)
+                add_and_overwrite(event, fields={target_field: source_fields_values}, rule=rule)
                 return
 
             case State(extend=True, overwrite=False, target_is_list=False):
                 source_fields_values = [target_field_value, *source_fields_values]
-                add_and_overwrite(event, target_field, source_fields_values)
+                add_and_overwrite(event, fields={target_field: source_fields_values}, rule=rule)
                 return
 
             case State(
@@ -151,21 +138,18 @@ class FieldManager(Processor):
             ):
                 flattened_source_fields = self._overwrite_from_source_values(source_fields_values)
                 source_fields_values = [*target_field_value, *flattened_source_fields]
-                add_and_overwrite(event, target_field, source_fields_values)
+                add_and_overwrite(event, fields={target_field: source_fields_values}, rule=rule)
                 return
 
             case State(overwrite=True, extend=True):
                 flattened_source_fields = self._overwrite_from_source_values(source_fields_values)
                 source_fields_values = [*flattened_source_fields]
-                add_and_overwrite(event, target_field, source_fields_values)
+                add_and_overwrite(event, fields={target_field: source_fields_values}, rule=rule)
                 return
 
             case _:
-                success = add_field_to(
-                    event, target_field, source_fields_values, state.extend, state.overwrite
-                )
-                if not success:
-                    raise FieldExistsWarning(rule, event, [target_field])
+                field = {target_field: source_fields_values}
+                add_fields_to(event, field, rule, state.extend, state.overwrite)
 
     def _overwrite_from_source_values(self, source_fields_values):
         duplicates = []